+++ src/main/java/nextstep/subway/station/ui/StationController.java
@@ -13,7 +13,7 @@ import java.util.List;
 
 @RestController
 public class StationController {
-    private StationService stationService;
+    private final StationService stationService;
 
     public StationController(StationService stationService) {
         this.stationService = stationService;
