+++ src/test/java/nextstep/subway/path/acceptance/PathAcceptanceTest.java
@@ -1,28 +1,20 @@
 package nextstep.subway.path.acceptance;
 
-import static nextstep.subway.line.acceptance.LineAcceptanceTest.지하철_노선_등록되어_있음;
-import static nextstep.subway.line.acceptance.LineSectionAcceptanceTest.지하철_노선에_지하철역_등록_요청;
-import static org.assertj.core.api.Assertions.assertThat;
+import static nextstep.subway.line.acceptance.LineAcceptanceSteps.지하철_노선_등록되어_있음;
+import static nextstep.subway.line.acceptance.LineSectionAcceptanceSteps.지하철_노선에_지하철역_등록_요청;
+import static nextstep.subway.path.acceptance.PathAcceptanceSteps.*;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 import nextstep.subway.AcceptanceTest;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
-import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.StationAcceptanceTest;
 import nextstep.subway.station.dto.StationResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 
 @DisplayName("지하철 경로 조회")
@@ -71,7 +63,7 @@ public class PathAcceptanceTest extends AcceptanceTest {
     }
 
     @Test
-    @DisplayName("출발역과 도착역이 같은 경우")
+    @DisplayName("출발역과 도착역이 같은 경우 {INTERNAL_SERVER_ERROR}")
     public void findPathSameSourceAndTarget() {
         // when
         ExtractableResponse<Response> response = 지하철_최단_경로_조회(교대역.getId(), 교대역.getId());
@@ -81,7 +73,7 @@ public class PathAcceptanceTest extends AcceptanceTest {
     }
 
     @Test
-    @DisplayName("출발역과 도착역이 연결이 되어 있지 않은 경우")
+    @DisplayName("출발역과 도착역이 연결이 되어 있지 않은 경우 {INTERNAL_SERVER_ERROR}")
     public void notConnectSourceAndTarget() {
         // given
         StationResponse 서울역 = StationAcceptanceTest.지하철역_등록되어_있음("서울역").as(StationResponse.class);
@@ -96,7 +88,7 @@ public class PathAcceptanceTest extends AcceptanceTest {
     }
 
     @Test
-    @DisplayName("존재하지 않은 출발역이나 도착역을 조회 할 경우")
+    @DisplayName("존재하지 않은 출발역이나 도착역을 조회 할 경우 {INTERNAL_SERVER_ERROR}")
     public void findPathNoExistStation() {
         // given
         Long 존재하지않는역 = 4885L;
@@ -106,46 +98,5 @@ public class PathAcceptanceTest extends AcceptanceTest {
 
         // then
         지하철_최단_경로_조회_실패됨(response);
-    }
-
-    public static ExtractableResponse<Response> 지하철_최단_경로_조회(Long sourceId, Long targetId) {
-        Map<String, String> params = new HashMap<>();
-        params.put("source", sourceId + "");
-        params.put("target", targetId + "");
-
-        return RestAssured
-            .given().log().all()
-            .queryParams(params)
-            .accept(MediaType.APPLICATION_JSON_VALUE)
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .when().get("/paths")
-            .then().log().all().extract();
-    }
-
-    public static void 지하철_최단_경로_지하철역_순서_정렬됨(ExtractableResponse<Response> response, List<StationResponse> expectedStations) {
-        PathResponse path = response.as(PathResponse.class);
-        List<Long> stationIds = path.getStations().stream()
-            .map(it -> it.getId())
-            .collect(Collectors.toList());
-
-        List<Long> expectedStationIds = expectedStations.stream()
-            .map(it -> it.getId())
-            .collect(Collectors.toList());
-
-        assertThat(stationIds).containsExactlyElementsOf(expectedStationIds);
-    }
-
-    public static void 지하철_최단_경로_거리(ExtractableResponse<Response> response, int distance) {
-        PathResponse path = response.as(PathResponse.class);
-
-        assertThat(path.getDistance()).isEqualTo(distance);
-    }
-
-    public static void 지하철_최단_경로_조회됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    public static void 지하철_최단_경로_조회_실패됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
     }
 }
